+++ packages/adapters/vue/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @splicetree/adapter-vue
 
+## 1.0.0
+
 ## 0.3.0
 
 ## 0.2.0