+++ packages/plugins/dnd/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @splicetree/plugin-dnd
 
+## 1.0.0
+
 ## 0.3.0
 
 ### Minor Changes